test.py was created for the purposes of directly running the driver commands, without needing to use ROS messages.
modificaitons in sphero.py allow the cmd_vel messages to correctly move the
Sphero, with a max velocity of 2.5 meters/second.  The Linear.x and Linear.y
are now interpretted correctly in moving the Sphero according to the right
hand coordinate system.  +X is forward, -X is reverse, +Y is left, -Y is right.